Hotkeys in dialogs. Step 3. EditLine keys
EditLine dialog control may have a hotkey. It reacts to the hotkey.
The hotkey is assigned to dependent StaticLabel element with the usual '&' in text.
One dialog: Commands-Associations-Ins is makred with the hotkeys to test the functionality. Everything appears to be working
Side effect. Originally, all EditLine inputs have treated Alt-letter as letter. This is needed for PanelSearchWin._edit. Now the acceptance of the Alt-keys is switchable on init. It is off by default, and is turned on for  PanelSearchWin._edit.

Minor change. Unix Shift-tab response in main loop ncwin.cpp is now consistent with Windows.

StaticLabel is derived from StaticText, which was used not only for element labels, but also to display various messages, including multi-line text, and system errors. Unlike the StaticText, StaticLabel is always single line of hard-coded content. It may have 'master' control, which will be activated on StaticLabel hotkey.

Remaining for the hotkeys
- Identify which other dialog elements may have hotkeys, and include the hotkeys there
- Markup all dialogs with hotkeys.
- Colors for hotkeys (red in default scheme on dialogs)

A new file, swl/swl_staticlabel.cpp is added to the project. Hope that I have updated all makefiles properly to reflect that.